Game canvas component
Вынес canvas в отдельный компонент
