+++ packages/client/src/shared/components/game-canvas/index.ts
@@ -1,0 +1,1 @@
+export * from './game-canvas'
