+++ packages/client/src/shared/components/index.ts
@@ -1,2 +1,3 @@
 export * from './error-boundary'
+export * from './game-canvas'
 export * from './app'
